+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Changes
 - Update to 1.21.10
+- Fixed default keybinding bug